Move simple-comment types to separate file
