+++ src/lib/blank-service.ts
@@ -1,46 +1,4 @@
-
-export type AuthToken = string
-export type CommentId = string
-export type DiscussionId = string
-export type Email = string
-export type URL = string
-export type UserId = string
-
-export interface Success {
-  code: number,
-  message: string
-}
-
-export interface Error {
-  code: number,
-  message: string
-}
-
-export interface Comment {
-  id: CommentId,
-  discussion: DiscussionId,
-  user: UserId,
-  text: string,
-  parentComment?: CommentId,
-  edits?:Comment[],
-  dateCreated: Date,
-  dateModified: Date
-}
-
-export interface User {
-  id:UserId,
-  verified:boolean,
-  email:Email,
-  name:string,
-  avatar:URL,
-  isAdmin:boolean
-}
-
-export interface Discussion {
-  id:DiscussionId,
-  isLocked: boolean,
-  comments: Comment[]
-}
+import type { AuthToken, CommentId, Discussion, DiscussionId, Success, User, UserId } from "./simple-comment";
 
 /**
  * Accept a user name and password, return authentication token
@@ -60,7 +18,7 @@ export const authPOST = () => new Promise<AuthToken | Error>((resolve, reject) =
  * commentId byte[] 
  * returns Success
  **/
-export const commentDELETE = (discussionId: DiscussionId, commentId: CommentId, authUser:UserId) => new Promise<Success | Error>((resolve, reject) => {
+export const commentDELETE = (discussionId: DiscussionId, commentId: CommentId, authUser: UserId) => new Promise<Success | Error>((resolve, reject) => {
   resolve()
 });
 
@@ -73,7 +31,7 @@ export const commentDELETE = (discussionId: DiscussionId, commentId: CommentId, 
  * commentId byte[] 
  * returns Comment
  **/
-export const commentGET = (discussionId: DiscussionId, commentId: CommentId, authUser?:UserId) => new Promise<Comment | Error>((resolve, reject) => {
+export const commentGET = (discussionId: DiscussionId, commentId: CommentId, authUser?: UserId) => new Promise<Comment | Error>((resolve, reject) => {
   resolve()
 });
 
@@ -86,7 +44,7 @@ export const commentGET = (discussionId: DiscussionId, commentId: CommentId, aut
  * commentId byte[] 
  * returns Comment
  **/
-export const discussionPUT = (discussionId: DiscussionId, commentId: CommentId, authUser:UserId) => new Promise<Comment | Error>((resolve, reject) => {
+export const discussionPUT = (discussionId: DiscussionId, commentId: CommentId, authUser: UserId) => new Promise<Comment | Error>((resolve, reject) => {
   resolve()
 });
 
@@ -98,7 +56,7 @@ export const discussionPUT = (discussionId: DiscussionId, commentId: CommentId, 
  * discussionId byte[] 
  * returns Success
  **/
-export const discussionDELETE = (discussionId: DiscussionId, authUser:UserId) => new Promise<Success | Error>((resolve, reject) => {
+export const discussionDELETE = (discussionId: DiscussionId, authUser: UserId) => new Promise<Success | Error>((resolve, reject) => {
   resolve()
 });
 
@@ -110,7 +68,7 @@ export const discussionDELETE = (discussionId: DiscussionId, authUser:UserId) =>
  * discussionId byte[] 
  * returns Discussion
  **/
-export const discussionGET = (discussionId: DiscussionId, authUser?:UserId) => new Promise<Discussion | Error>((resolve, reject) => {
+export const discussionGET = (discussionId: DiscussionId, authUser?: UserId) => new Promise<Discussion | Error>((resolve, reject) => {
   resolve()
 });
 
@@ -122,7 +80,7 @@ export const discussionGET = (discussionId: DiscussionId, authUser?:UserId) => n
  * discussionId byte[] 
  * returns Comment
  **/
-export const discussionPOST = (discussionId: DiscussionId, authUser:UserId) => new Promise<Comment | Error>((resolve, reject) => {
+export const discussionPOST = (discussionId: DiscussionId, authUser: UserId) => new Promise<Comment | Error>((resolve, reject) => {
   resolve()
 });
 
@@ -134,7 +92,7 @@ export const discussionPOST = (discussionId: DiscussionId, authUser:UserId) => n
  *
  * returns List
  **/
-export const discussionListGET = (authUser?:UserId) => new Promise<Discussion[] | Error>((resolve, reject) => {
+export const discussionListGET = (authUser?: UserId) => new Promise<Discussion[] | Error>((resolve, reject) => {
   resolve()
 });
 
@@ -145,7 +103,7 @@ export const discussionListGET = (authUser?:UserId) => new Promise<Discussion[] 
  *
  * returns List
  **/
-export const userListGET = (authUser?:UserId) => new Promise<User[] | Error>((resolve, reject) => {
+export const userListGET = (authUser?: UserId) => new Promise<User[] | Error>((resolve, reject) => {
   resolve()
 });
 
@@ -167,7 +125,7 @@ export const userPOST = () => new Promise<User | Error>((resolve, reject) => {
  * userId byte[] 
  * returns Success
  **/
-export const userDELETE = (userId: UserId, authUser:UserId) => new Promise<Success | Error>((resolve, reject) => {
+export const userDELETE = (userId: UserId, authUser: UserId) => new Promise<Success | Error>((resolve, reject) => {
   resolve()
 });
 
@@ -179,7 +137,7 @@ export const userDELETE = (userId: UserId, authUser:UserId) => new Promise<Succe
  * userId byte[] 
  * returns User
  **/
-export const userGET = (userId: UserId, authUser?:UserId) => new Promise<User | Error>((resolve, reject) => {
+export const userGET = (userId: UserId, authUser?: UserId) => new Promise<User | Error>((resolve, reject) => {
   resolve()
 });
 
@@ -191,7 +149,7 @@ export const userGET = (userId: UserId, authUser?:UserId) => new Promise<User | 
  * userId byte[] 
  * returns Success
  **/
-export const userPUT = (userId: UserId, authUser:UserId) => new Promise<Success | Error>((resolve, reject) => {
+export const userPUT = (userId: UserId, authUser: UserId) => new Promise<Success | Error>((resolve, reject) => {
   resolve()
 });
